DateCalendar component builds a month calendar (#3)
Previously, the DateCalendar component held temporary buttons for
changing the sliding pages. This commit removes that temporary code
and implements the actual component. The DateCalendar component has a
required month and year parameter, and creates a month calendar for
that month. The DateCalendar component also accepts an importantDates
parameter with a corresponding onClick function that will create
clickable dates in the month calendar. When an important date link is
clicked, that date will be broadcast to the onClick function.

This component isn't fully completed as the imporantDate links will be
improved in a later commit.

Updated example useage of this component:
```
<DateCalendar
  month={0-11}
  year={2018}
  importantDates=[{ date: '2018-06-09T00:00:50Z' }]
  onClick={onClickFunction}
/>
```

Additionally, thic commit implements the DateCalender component to
create monthly calendars for May and June, 2018 with a few possible
imporant dates for our wedding.